Update: clearing post user relation when deleting

+++ FirstAPI/controllers/feed.js
@@ -128,7 +128,12 @@ exports.updatePost = (req, res, next) => {
     .then(post => {
       if (!post) {
         const error = new Error('Could not find post!');
-        error.statusCode(404);
+        error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
         throw error;
       }
       if (imageUrl !== post.imageUrl) {
@@ -160,12 +165,22 @@ exports.deletePost = (req, res, next) => {
         error.statusCode(404);
         throw error;
       }
-      //check logged user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
     })
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       return res.status(200).json({ message: 'Post deleted successfully!' })
     })
     .catch(err => {
